refactor: memoized sortedGroupedStakes and created a component for the cb fn on Dashboard
